Se agrego autenticacion a la aplicacion usando auth0

+++ app/Http/Controllers/ReservationController.php
@@ -19,6 +19,7 @@ class ReservationController extends Controller
     public function index()
     {
         $reservations = Reservation::with('room', 'room.hotel')
+        ->where('user_id', \Auth::user()->getUserInfo()['sub'])
         ->orderBy('arrival', 'asc')
         ->get();
 
@@ -30,7 +31,7 @@ class ReservationController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create($hotel_id)
     {
         $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
         return view('dashboard.reservationCreate', compact('hotelInfo'));
@@ -44,7 +45,12 @@ class ReservationController extends Controller
      */
     public function store(Request $request) /// agregar validaciones
     {
-        $request->request->add(['user_id' => 1]);
+        // Set the user_id equal to the user's Auth0 sub id before
+        // Will be similar to "auth0|123123123123123"
+        $user_id = \Auth::user()->getUserInfo()['sub'];
+        $request->request->add(['user_id' => $user_id]);
+        
+        // Create the request
         Reservation::create($request->all());
 
         return redirect('dashboard/reservations')->with('success', 'Reservation created!');
@@ -56,13 +62,21 @@ class ReservationController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Reservation $reservation)
     {
         $reservation = Reservation::with('room', 'room.hotel')->get()->find($reservation->id);
-        $hotel_id = $reservation->room->hotel_id;
-        $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
 
-        return view('dashboard.reservationSingle', compact('reservation', 'hotelInfo'));
+        if($reservation->user_id === \Auth::user()->getUserInfo()['sub'])
+        {
+            $hotel_id = $reservation->room->hotel_id;
+            $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
+
+            return view('dashboard.reservationSingle', compact('reservation', 'hotelInfo'));
+        }
+        else
+        {
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to see that.');
+        }
     }
 
     /**
@@ -71,13 +85,21 @@ class ReservationController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Reservation $reservation)
     {
         $reservation = Reservation::with('room', 'room.hotel')->get()->find($reservation->id);
-        $hotel_id = $reservation->room->hotel_id;
-        $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
 
-        return view('dashboard.reservationEdit', compact('reservation', 'hotelInfo'));
+        if ($reservation->user_id === \Auth::user()->getUserInfo()['sub'])
+        {
+            $hotel_id = $reservation->room->hotel_id;
+            $hotelInfo = Hotel::with('rooms')->get()->find($hotel_id);
+
+            return view('dashboard.reservationEdit', compact('reservation', 'hotelInfo'));
+        }
+        else
+        {
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to do that');
+        }
     }
 
     /**
@@ -87,9 +109,17 @@ class ReservationController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Reservation $reservation)
     {
-        $reservation->user_id = 1;
+        if ($reservation->user_id != \Auth::user()->getUserInfo()['sub'])
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to update this reservation');
+
+        $user_id = \Auth::user()->getUserInfo()['sub'];
+        $reservation->user_id = $user_id;
+        $reservation->num_of_guests = $request->num_of_guests;
+        $reservation->arrival = $request->arrival;
+        $reservation->departure = $request->departure;
+        $reservation->room_id = $request->room_id;
 
         $reservation->save();
         return redirect('dashboard/reservations')->with('success', 'Successfully updated your reservation!');
@@ -101,11 +131,19 @@ class ReservationController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id) /// agregar validaciones
+    public function destroy(Reservation $reservation) /// agregar validaciones
     {
         $reservation = Reservation::find($reservation->id);
-        $reservation->delete(); 
 
-        return redirect('dashboard/reservations')->with('success', 'Successfully deleted your reservation!');
+        if ($reservation->user_id === \Auth::user()->getUserInfo()['sub'])
+        {
+            $reservation->delete(); 
+
+            return redirect('dashboard/reservations')->with('success', 'Successfully deleted your reservation!');
+        }
+        else
+        {
+            return redirect('dashboard/reservations')->with('error', 'You are not authorized to delete this reservation');
+        }
     }
 }
